Fixed it so tests will pass
