+++ lib/storage.js
@@ -260,6 +260,7 @@ File.prototype._checkDone = function () {
   })
 
   if (self.done) {
+    debug('done');
     process.nextTick(function () {
       self.emit('done')
     })
@@ -550,6 +551,7 @@ Storage.prototype._checkDone = function () {
 
   if (!self.done && self.files.every(function (file) { return file.done })) {
     self.done = true
+    debug('done');
     self.emit('done')
   }
 }
