+++ lib/torrent-manager.js
@@ -7,12 +7,19 @@ var parseTorrent = require('parse-torrent')
 var Torrent = require('./torrent')
 var Storage = require('./storage')
 var Server = require('./server') // browser exclude
+var hh = require('http-https') // browser exclude
+var once = require('once')
+var concat = require('concat-stream') // browser exclude
+var reemit = require('re-emitter')
+var parallel = require('run-parallel')
+var fs = require('fs') // browser exclude
 
 inherits(torrentManager, EventEmitter)
 
  function torrentManager(torrentId, opts) {
   var self = this
- self._storageImpl = opts.storage || Storage
+ self._storageImpl = opts.storage || Storage 
+ self.files = []
 
   var parsedTorrent = parseTorrent(torrentId)
   if (parsedTorrent && parsedTorrent.infoHash) {
@@ -42,14 +49,21 @@ inherits(torrentManager, EventEmitter)
     debug(self.infoHash);
     if (parsedTorrent.name) self.name = parsedTorrent.name // preliminary name
     self.torrent = new Torrent(self, parsedTorrent, opts);
+    reemit(self.torrent, self, ['ready', 'dhtAnnounce', 'listening']);
+    self.torrent.on('ready', function() { self.ready = true; });
     if (parsedTorrent.info) self.onMetadata(parsedTorrent)
   }
+}
+
+torrentManager.prototype.addPeer = function(peer) {
+  var self = this
+  self.torrent.addPeer(peer);
 }
 
 torrentManager.prototype.createServer = function (opts) {
   var self = this
   if (typeof Server === 'function' /* browser exclude */) {
-    return new Server(self.torrent, opts)
+    return new Server(self, opts)
   }
 }
 
@@ -75,6 +89,15 @@ torrentManager.prototype.onMetadata = function (metadata) {
   }
 
   self.storage = new self._storageImpl(self.parsedTorrent, self.storageOpts)
+  self.storage.files.forEach(function (file) {
+    self.files.push(file)
+  })
+  
+  self.storage.on('done', function () {
+    debug('torrent ' + self.infoHash + ' done')
+    self.emit('done')
+  })
+  
   if (self.verify) {
     process.nextTick(function () {
       debug('verifying existing torrent data')
@@ -111,4 +134,52 @@ torrentManager.prototype.onMetadata = function (metadata) {
     self.torrent.updateMetadata(self.parsedTorrent)
     self.emit('metadata')
   })
+}
+
+/**
+ * Destroy and cleanup this torrent.
+ */
+torrentManager.prototype.destroy = function (cb) {
+  var self = this
+  debug('destroy')
+  self._destroyed = true
+  clearInterval(self._rechokeIntervalId)
+
+  var tasks = []
+  if (self.torrent) tasks.push(function (cb) {
+    self.torrent.destroy(cb)
+  })
+
+  parallel(tasks, cb)
+}
+
+
+/**
+ * Make http or https request, following redirects.
+ * @param {string} u
+ * @param {function}
+ * @param {number=} maxRedirects
+ * @return {http.ClientRequest}
+ */
+function httpGet (u, cb, maxRedirects) {
+  cb = once(cb)
+  if (!maxRedirects) maxRedirects = 5
+  if (maxRedirects === 0) return cb(new Error('too many redirects'))
+
+  hh.get(u, function (res) {
+    // Check for redirect
+    if (res.statusCode >= 300 && res.statusCode < 400 && 'location' in res.headers) {
+      var location = res.headers.location
+      if (!url.parse(location).host) {
+        // If relative redirect, prepend host of current url
+        var parsed = url.parse(u)
+        location = parsed.protocol + '//' + parsed.host + location
+      }
+      res.resume() // discard response
+      return httpGet(location, cb, --maxRedirects)
+    }
+    res.pipe(concat(function (data) {
+      cb(null, data)
+    }))
+  }).on('error', cb)
 }
