+++ lib/torrent.js
@@ -59,7 +59,6 @@ function Torrent (manager, parsedTorrent, opts) {
   self._rechokeIntervalId = null
 
   self.ready = false
-  self.files = []
   self.metadata = null
   self.parsedTorrent = parsedTorrent
   self.infoHash = parsedTorrent.infoHash
@@ -154,17 +153,11 @@ Torrent.prototype.attachStorage = function(storage) {
       self.discovery.tracker.complete()
 
     debug('torrent ' + self.infoHash + ' done')
-    self.emit('done')
   })
 
   self.storage.on('select', self.select.bind(self))
   self.storage.on('deselect', self.deselect.bind(self))
   self.storage.on('critical', self.critical.bind(self))
-
-  self.storage.files.forEach(function (file) {
-    self.files.push(file)
-  })
-  
   process.nextTick(self._onStorage.bind(self))
 }
 
@@ -176,7 +169,11 @@ Torrent.prototype.updateMetadata = function(parsedTorrent) {
   self.metadata = parseTorrent.toBuffer(parsedTorrent)
 
   // update discovery module with full torrent metadata
-  self.discovery.setTorrent(self.parsedTorrent)
+  if(!!self.discovery) {
+    self.discovery.setTorrent(self.parsedTorrent)
+  } else {
+    self.on('listening', function() { self.discovery.setTorrent(self.parsedTorrent) });
+  }
 
   self.rarityMap = new RarityMap(self.swarm, self.parsedTorrent.pieces.length)
 
@@ -912,34 +909,4 @@ function randomizedForEach (array, cb) {
   indices.forEach(function (index) {
     cb(array[index], index, array)
   })
-}
-
-/**
- * Make http or https request, following redirects.
- * @param {string} u
- * @param {function}
- * @param {number=} maxRedirects
- * @return {http.ClientRequest}
- */
-function httpGet (u, cb, maxRedirects) {
-  cb = once(cb)
-  if (!maxRedirects) maxRedirects = 5
-  if (maxRedirects === 0) return cb(new Error('too many redirects'))
-
-  hh.get(u, function (res) {
-    // Check for redirect
-    if (res.statusCode >= 300 && res.statusCode < 400 && 'location' in res.headers) {
-      var location = res.headers.location
-      if (!url.parse(location).host) {
-        // If relative redirect, prepend host of current url
-        var parsed = url.parse(u)
-        location = parsed.protocol + '//' + parsed.host + location
-      }
-      res.resume() // discard response
-      return httpGet(location, cb, --maxRedirects)
-    }
-    res.pipe(concat(function (data) {
-      cb(null, data)
-    }))
-  }).on('error', cb)
 }
